+++ bot/types/bot.go
@@ -7,7 +7,7 @@ type Bot struct {
 }
 
 type Message struct {
-	ChatId int    `xml:"chat_id"`
+	ChatId int64  `xml:"chat_id"`
 	Data   string `xml:"text"`
 }
 
@@ -21,7 +21,7 @@ type TelegramUpdate struct {
 	Message  struct {
 		MessageID int `json:"message_id"`
 		From      struct {
-			ID           int    `json:"id"`
+			ID           int64  `json:"id"`
 			IsBot        bool   `json:"is_bot"`
 			FirstName    string `json:"first_name"`
 			LastName     string `json:"last_name"`
@@ -29,7 +29,7 @@ type TelegramUpdate struct {
 			LanguageCode string `json:"language_code"`
 		} `json:"from"`
 		Chat struct {
-			ID        int    `json:"id"`
+			ID        int64  `json:"id"`
 			FirstName string `json:"first_name"`
 			LastName  string `json:"last_name"`
 			Username  string `json:"username"`
@@ -47,7 +47,7 @@ type TelegramUpdate struct {
 			MimeType     string `json:"mime_type"`
 			FileID       string `json:"file_id"`
 			FileUniqueID string `json:"file_unique_id"`
-			FileSize     int    `json:"file_size"`
+			FileSize     int64  `json:"file_size"`
 		} `json:"document"`
 	} `json:"message"`
 }
